improve .gitconfig update workflow

+++ updateConfig.sh
@@ -1,8 +1,11 @@
 #!/bin/bash
 
-echo; echo "Copying git config files...";
-cp ~/.dotfiles/config/.gitconfig ~/.gitconfig
-echo "~/.gitconfig updated!"
+if [[ ! -f ~/.gitconfig ]]
+then
+    echo; echo "Copying git config files...";
+    cp ~/.dotfiles/config/.gitconfig ~/.gitconfig
+    echo "~/.gitconfig updated!"
+fi
 
 echo "Copying zsh config files...";
 cp ~/.dotfiles/config/.zshrc ~/.zshrc
